hotfix: alteração no tipo da variável do método 'encontrarPorId'

+++ src/main/java/br/com/invistaix/InvistaIX/api/EnderecoRestController.java
@@ -32,7 +32,7 @@ public class EnderecoRestController {
 	}
 	
 	@GetMapping("/findById={id}")
-	public EnderecoModel encontrarPorId(@PathVariable Integer id) {
+	public EnderecoModel encontrarPorId(@PathVariable Long id) {
 		try {
 			return enderecoService.encontrarPorId(id);
 		} catch (IllegalArgumentException ex) {
